+++ config/runtime.exs
@@ -16,6 +16,13 @@ import Config
 #
 # Alternatively, you can use `mix phx.gen.release` to generate a `bin/server`
 # script that automatically sets the env var above.
+
+empty_or_nil? = fn
+  {_key, nil} -> true
+  {_key, ""} -> true
+  _value -> false
+end
+
 if System.get_env("PHX_SERVER") do
   config :td_audit, TdAuditWeb.Endpoint, server: true
 end
@@ -72,21 +79,27 @@ if config_env() == :prod do
         optional_db_ssl_options_keyfile
 
   email_tls_versions =
-    "SMTP_TLS_VERSIONS"
-    |> System.get_env("tlsv1.2")
+    System.get_env("SMTP_TLS_VERSIONS", "tlsv1.2")
     |> String.split(",")
     |> Enum.map(&String.to_atom/1)
 
-  config :td_audit, TdAudit.Notifications.Mailer,
-    server: System.get_env("SMTP_SERVER"),
-    port: System.get_env("SMTP_PORT"),
-    username: System.get_env("SMTP_USERNAME"),
-    password: System.get_env("SMTP_PASSWORD"),
-    tls: System.get_env("SMTP_TLS", "always") |> String.to_atom(),
-    allowed_tls_versions: email_tls_versions,
-    tls_verify: System.get_env("SMTP_TLS_VERIFY", "verify_none") |> String.to_atom(),
-    ssl: false,
-    retries: 3
+  smtp_auth =
+    Enum.reject(
+      [username: System.get_env("SMTP_USERNAME"), password: System.get_env("SMTP_PASSWORD")],
+      &empty_or_nil?.(&1)
+    )
+
+  config :td_audit,
+         TdAudit.Notifications.Mailer,
+         [
+           server: System.get_env("SMTP_SERVER"),
+           port: System.get_env("SMTP_PORT"),
+           tls: System.get_env("SMTP_TLS", "always") |> String.to_atom(),
+           allowed_tls_versions: email_tls_versions,
+           tls_verify: System.get_env("SMTP_TLS_VERIFY", "verify_none") |> String.to_atom(),
+           ssl: false,
+           retries: 3
+         ] ++ smtp_auth
 
   config :td_cache,
     redis_host: System.fetch_env!("REDIS_HOST"),